Create class GmlasPanelMixin
